pyproject.toml
[project]
name = "devops-mcp"
version = "0.1.0"
description = "Local dev co-pilot MCP server — inspect Docker containers, processes, system info, and run sandboxed shell commands"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docker>=7.1.0",
    "mcp[cli]>=1.26.0",
    "psutil>=7.2.2",
    "python-dotenv>=1.2.2",
]

[project.scripts]
devops-mcp = "devops_mcp.server:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/devops_mcp"]
